requirements.txt
aldryn-bootstrap3==1.0.10
aldryn-client==2.3.1
aldryn-style==1.1.1
click==6.6
dj-database-url==0.4.1
Django==1.9.7
django-appconf==1.0.2
django-classy-tags==0.7.2
django-cms==3.3.0
django-durationfield==0.5.2
django-filer==1.2.0
django-formtools==1.0
django-mptt==0.8.4
django-polymorphic==0.8.1
django-reversion==1.10.2
django-sekizai==0.9.0
Django-Select2==4.3.2
django-treebeard==4.0.1
djangocms-admin-style==1.2.2
djangocms-column==1.6.0
djangocms-file==1.0
djangocms-googlemap==0.4.0
djangocms-inherit==0.2.1
djangocms-installer==0.8.10
djangocms-link==1.7.2
djangocms-picture==0.2.0
djangocms-style==1.7.0
djangocms-teaser==0.2.0
djangocms-text-ckeditor==3.0.0
djangocms-video==1.0.0
easy-thumbnails==2.3
html5lib==0.9999999
Pillow==3.2.0
pytz==2016.4
requests==2.10.0
six==1.10.0
smart-mirror-clock==0.0.3
smart-mirror-weather==0.0.3
tabulate==0.7.5
tzlocal==1.2.2
Unidecode==0.4.19
wheel==0.26.0
